Fix id generation in NumberField and TextField (#92)
* docs(text-field): fix story

* fix(number-field): ensure id

Previously it could happen that no id was generated when id={undefined} was passed explicitly. The tests were not catching that and were upgraded as well.

* fix(text-field): ensure id

Previously it could happen that no id was generated when id={undefined} was passed explicitly. The tests were not catching that and were upgraded as well.