feat(money-field): mark both fields as touched when one of them loses focus
We only mark both as blurred when an element which is not part of the MoneyInput gains focus. This is done to ensure the MoneyInput acts as one unit so that errors can be displayed correctly (as they rely on both touched states and would pop up too early if only one of them is checked).

This also auto-focuses the amount intput after a currency has been selected.