Redo MoneyInput (better UX, less error cases) (#175)
* feat(money-input): reduce to hasError and hasWarning

BREAKING CHANGE: dropped support for hasAmountError, hasAmountWarning, hasCurrencyError and hasCurrencyWarning. Use hasError and hasWarning instead.

BREAKING CHANGE: isTouched only returns true when both fields were touched from now on.

* feat(money-field): mark both fields as touched when one of them loses focus

We only mark both as blurred when an element which is not part of the MoneyInput gains focus. This is done to ensure the MoneyInput acts as one unit so that errors can be displayed correctly (as they rely on both touched states and would pop up too early if only one of them is checked).

This also auto-focuses the amount intput after a currency has been selected.

* test(money-input): switch tests to react-testing-library

* docs(money-input): fix leftover areAllTouched