Redo TimeInput (#151)
* test(test-utils): add abilitiy to set locale

* feat(time-input): redo component

Reimagine the component and drop flatpickr.

* test(time-input): work around Intl API

Workaround for the node Intl API (not react-intl itself).

Had to stop using props.intl.formatTime as it was not respecting the passed locale when formatting time.

* feat(time-input): support high precision

* refactor(time-input): simplify high precision handling

* docs(time-input): explain form usage

* feat(time-input): make clock icon a label

And change cursor to default one for the input.

* test(time-input): group TimeInput.to24h tests

* style(time-input): remove outline

Yeah, I know it's bad. The other inputs do not use an outline either right now. There is alternative styling instead. We are discussiong to readd outlines at once eventually.

* fix(time-input): fix milliseconds parsing

* refactor(time-input): use default export for time-input-body

* fix(time-input): avoid 12am/12pm jump when blurring repeatedly

* fix(time-input): format time when blurred

* fix(time-input): do not allow hours>12 for 12h clock