refactor: to use `@babel/preset-typescript` for TypeScript transpilation (#457)
#### Summary

This pull request suggests/evaluates using the `@babel/preset-typescript` to compile out TypeScript.

#### Description

`@babel/preset-typescript` comes with most of (not all) TypeScript features to be compiled away. Generally, "How does Babel handle TypeScript code? It removes it". It supports most features. I only know that it does not not support namespaces or const enums because those require type information to transpile. Also does not support export = and import =, because those cannot be transpiled to ESM. I would argue that this could be fine as the initial step. Furthermore, babel is known to be faster than tsc (totally polemic argument though given the size of this code base).

If we would want to go this way we should add an `yarn type-check` script that invokes the TypeScript compiler. As this changes the workflow a bit - allowing easier experimentation - as you write code, you save, and it compiles (very quickly) without checking for type safety just yet.

1. Generally I like that we say within babel for both webpack and rollup. 
   - It's not like we sent our files through TS and then Babel. 
2. We can continue to use any JavaScript proposal and syntax we like (requires some thought)
   - Do we use `private` or `#instanceVar`?
   - do expressions could be a thing (not that they should)